requirements.txt
requests>=2.31.0
pillow>=10.0.0
pystray>=0.19.0
plyer>=2.1.0
pyinstaller>=6.0.0
